+++ test/pull_submodules/test_pull_submodules.py
@@ -38,12 +38,26 @@ def mkdir_app(directory, resolve_json):
         rename_as="resolve.json",
     )
 
-    app_dir.copy_file("test/add_dependency/fake_git_clone.py")
+    app_dir.copy_file("test/fake_git.py")
     app_dir.copy_file("build/waf")
 
     # Note: waf will call "git config --get remote.origin.url" in this folder,
     # so "git init" is required to test the default behavior (https resolver)
-    app_dir.run(["git", "init"])
+    # app_dir.run(["git", "init"])
+    git_info = {
+        "checkout": "master",
+        "branches": ["master"],
+        "commits": [],
+        "remote_origin_url": "git@github.com/acme-corp/app.git",
+        "is_detached_head": False,
+        "submodules": [],
+        "tags": [],
+    }
+
+    json_path = os.path.join(app_dir.path(), "git_info.json")
+
+    with open(json_path, "w") as json_file:
+        json.dump(git_info, json_file)
 
     return app_dir
 
@@ -66,39 +80,46 @@ def commit_file(directory, filename, content):
 
 
 def mkdir_libqux(directory):
-    # Add bar dir
+    # Add qux dir
     qux_dir = directory.mkdir("libqux")
-    qux_dir.run(["git", "init"])
-    commit_file(directory=qux_dir, filename="ok.txt", content="hello world")
+    qux_dir.write_text(filename="ok.txt", data="hello world", encoding="utf-8")
+
+    git_info = {
+        "checkout": "master",
+        "branches": ["master"],
+        "commits": [],
+        "remote_origin_url": "git@github.com/acme/qux.git",
+        "is_detached_head": False,
+        "submodules": [],
+        "tags": [],
+    }
+
+    json_path = os.path.join(qux_dir.path(), "git_info.json")
+
+    with open(json_path, "w") as json_file:
+        json.dump(git_info, json_file)
 
     return qux_dir
 
 
 def mkdir_libbaz(directory, qux_dir):
     # Add baz dir
-    baz_dir = directory.copy_dir(directory="test/pull_submodules/libbaz")
-    baz_dir.run(["git", "init"])
-    baz_dir.run(["git", "add", "."])
-    baz_dir.run(["git", "submodule", "add", qux_dir.path(), "libqux"])
-    baz_dir.run(["git", "add", "."])
-    baz_dir.run(
-        [
-            "git",
-            "-c",
-            "user.name=John",
-            "-c",
-            "user.email=doe@email.org",
-            "commit",
-            "-m",
-            "oki",
-        ]
-    )
-    baz_dir.run(["git", "tag", "3.1.2"])
+    baz_dir = directory.copy_dir(directory="test/add_dependency/libbaz")
 
-    # We now remove the submodule dir - if pull_submodules is true in
-    # resolve.json it will get recreated.
-    libqux = baz_dir.join("libqux")
-    libqux.rmdir()
+    git_info = {
+        "checkout": "master",
+        "branches": ["master"],
+        "commits": [],
+        "remote_origin_url": "git@github.com/acme/baz.git",
+        "is_detached_head": False,
+        "submodules": [("libqux", qux_dir.path())],
+        "tags": ["3.1.2"],
+    }
+
+    json_path = os.path.join(baz_dir.path(), "git_info.json")
+
+    with open(json_path, "w") as json_file:
+        json.dump(git_info, json_file)
 
     return baz_dir
 
@@ -119,7 +140,7 @@ def test_pull_submodule(testdirectory):
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
-    clone_path = {"gitlab.com/acme/baz.git": baz_dir.path()}
+    clone_path = {"acme/baz.git": baz_dir.path()}
 
     json_path = os.path.join(app_dir.path(), "clone_path.json")
 
@@ -127,7 +148,9 @@ def test_pull_submodule(testdirectory):
         json.dump(clone_path, json_file)
 
     # Try the use checkout
-    app_dir.run(["python", "waf", "configure", "-v"])
+    app_dir.run(
+        ["python", "waf", "configure", "-v", "--resolve_path", "resolved_dependencies"]
+    )
     app_dir.run(["python", "waf", "build", "-v"])
 
     assert os.path.exists(os.path.join(app_dir.path(), "resolve_symlinks", "baz"))
@@ -153,7 +176,7 @@ def test_no_pull_submodule(testdirectory):
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
-    clone_path = {"gitlab.com/acme/baz.git": baz_dir.path()}
+    clone_path = {"acme/baz.git": baz_dir.path()}
 
     json_path = os.path.join(app_dir.path(), "clone_path.json")