+++ test/python/test_virtualenv.py
@@ -37,32 +37,34 @@ def test_virtualenv_name(testdirectory):
     assert testdirectory.contains_dir(name)
 
     venv = VirtualEnv.create(
-        log=log,
-        cwd=cwd,
-        env=env,
-        name=name,
-        ctx=ctx,
-        system_site_packages=True,
-        download=False,
+        log=log, cwd=cwd, env=env, name=name, ctx=ctx, system_site_packages=True
     )
 
     assert fnmatch.fnmatch(venv.path, os.path.join(cwd, name))
     assert not testdirectory.contains_dir(name)
 
-    ctx.cmd_and_log.assert_called_once_with(
-        [sys.executable, "-m", "virtualenv", name, "--system-site-packages"],
-        cwd=cwd,
-        env=env,
+    ctx.cmd_and_log.assert_has_calls(
+        [
+            mock.call(
+                [sys.executable, "-m", "venv", name, "--system-site-packages"],
+                cwd=venv.cwd,
+                env=venv.env,
+            )
+        ]
     )
 
     venv.run("python -m pip install pytest")
 
-    ctx.exec_command.assert_called_once_with(
-        "python -m pip install pytest",
-        cwd=venv.cwd,
-        env=venv.env,
-        stdout=None,
-        stderr=None,
+    ctx.exec_command.assert_has_calls(
+        [
+            mock.call(
+                "python -m pip install pytest",
+                cwd=venv.cwd,
+                env=venv.env,
+                stdout=None,
+                stderr=None,
+            ),
+        ]
     )
 
 
@@ -80,28 +82,27 @@ def test_virtualenv_system_site_packages(testdirectory):
     assert testdirectory.contains_dir(name)
 
     venv = VirtualEnv.create(
-        log=log,
-        cwd=cwd,
-        env=env,
-        name=name,
-        ctx=ctx,
-        system_site_packages=False,
-        download=False,
+        log=log, cwd=cwd, env=env, name=name, ctx=ctx, system_site_packages=False
     )
 
     assert fnmatch.fnmatch(venv.path, os.path.join(cwd, name))
     assert not testdirectory.contains_dir(name)
 
-    ctx.cmd_and_log.assert_called_once_with(
-        [sys.executable, "-m", "virtualenv", name], cwd=cwd, env=env
+    ctx.cmd_and_log.assert_has_calls(
+        [mock.call([sys.executable, "-m", "venv", name], cwd=venv.cwd, env=venv.env)]
     )
 
     venv.run("python -m pip install pytest")
 
-    ctx.exec_command.assert_called_once_with(
-        "python -m pip install pytest",
-        cwd=venv.cwd,
-        env=venv.env,
-        stdout=None,
-        stderr=None,
+    ctx.exec_command.assert_has_calls(
+        [
+            # Either "python -m ensurepip" or "python get-pip.py"
+            mock.call(
+                "python -m pip install pytest",
+                cwd=venv.cwd,
+                env=venv.env,
+                stdout=None,
+                stderr=None,
+            ),
+        ]
     )